# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Basis Lab API",
    "version": "0.1.0",
    "description": "Read-only normalized venue OHLC and cross-venue premium/discount candles. Timestamps are Unix milliseconds. No authentication is required.",
    "license": { "name": "MIT", "identifier": "MIT" }
  },
  "servers": [{ "url": "/", "description": "Current Rust deployment" }],
  "security": [],
...
  "paths": {
...
    "/api/v1/markets": {
      "get": {
        "operationId": "listMarkets",
        "summary": "Search a venue's cached native and normalized market symbols",
        "tags": ["Market data"],
        "parameters": [
          { "$ref": "#/components/parameters/Venue" },
          { "name": "query", "in": "query", "description": "Case-insensitive native or normalized ticker search; separators are optional", "schema": { "type": "string", "default": "", "maxLength": 64 }, "example": "BTC/USDT" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 1000, "default": 200 } }
        ],
        "responses": {
          "200": {
            "description": "Matching active markets",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MarketsResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "502": { "$ref": "#/components/responses/UpstreamError" },
          "503": { "$ref": "#/components/responses/Busy" },
          "504": { "$ref": "#/components/responses/UpstreamTimeout" }
        }
      }
    },
    "/api/v1/tickers": {
      "get": {
        "operationId": "searchTickers",
        "summary": "Search the cached normalized ticker catalog across venues",
        "description": "Returns venue-native symbols alongside canonical BASE/QUOTE identities. The bounded catalog is coalesced and cached for five minutes.",
        "tags": ["Market data"],
        "parameters": [
          { "name": "venue", "in": "query", "description": "Optional venue filter", "schema": { "$ref": "#/components/schemas/Venue" } },
          { "name": "query", "in": "query", "description": "Case-insensitive ticker, asset, venue, or native-symbol search; separators are optional", "schema": { "type": "string", "default": "", "maxLength": 64 }, "example": "WLFI/USDT" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 1000, "default": 200 } }
        ],
        "responses": {
          "200": {
            "description": "Matching cached ticker listings ranked by relevance",
            "headers": { "Cache-Control": { "schema": { "type": "string" }, "description": "Public browser cache and stale-while-revalidate policy" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TickersResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "502": { "$ref": "#/components/responses/UpstreamError" },
          "503": { "$ref": "#/components/responses/Busy" },
          "504": { "$ref": "#/components/responses/UpstreamTimeout" }
        }
      }
    },
    "/api/v1/candles": {
      "get": {
        "operationId": "getCandles",
        "summary": "Get normalized OHLC candles",
        "description": "Returns ascending candles in the venue's native market notation. Source bars are bucket-aligned, validated, deduplicated, and range-limited.",
        "tags": ["Market data"],
        "parameters": [
          { "$ref": "#/components/parameters/Venue" },
          { "$ref": "#/components/parameters/Market" },
          { "$ref": "#/components/parameters/Interval" },
          { "$ref": "#/components/parameters/From" },
          { "$ref": "#/components/parameters/To" },
          { "$ref": "#/components/parameters/Limit" }
        ],
        "responses": {
          "200": {
            "description": "Normalized source candles",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CandleResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "502": { "$ref": "#/components/responses/UpstreamError" },
          "503": { "$ref": "#/components/responses/Busy" },
          "504": { "$ref": "#/components/responses/UpstreamTimeout" }
        }
      }
    },
    "/api/v1/compare": {
      "get": {
        "operationId": "compareMarkets",
        "summary": "Calculate cross-venue premium/discount candles",
        "description": "Calculates (left / right - 1) × scale. Open and close use like-for-like fields. High uses left.high / right.low and low uses left.low / right.high, a conservative envelope because OHLC cannot prove simultaneous venue extremes. Only identical normalized opening timestamps are joined.",
        "tags": ["Comparison"],
        "parameters": [
          { "name": "left_venue", "in": "query", "required": true, "schema": { "$ref": "#/components/schemas/Venue" } },
          { "name": "left_market", "in": "query", "required": true, "schema": { "$ref": "#/components/schemas/MarketSymbol" }, "example": "WLFIUSDT" },
          { "name": "right_venue", "in": "query", "required": true, "schema": { "$ref": "#/components/schemas/Venue" } },
          { "name": "right_market", "in": "query", "required": true, "schema": { "$ref": "#/components/schemas/MarketSymbol" }, "example": "WLFI_USDT" },
          { "$ref": "#/components/parameters/Interval" },
          { "$ref": "#/components/parameters/From" },
          { "$ref": "#/components/parameters/To" },
          { "$ref": "#/components/parameters/Limit" },
          { "name": "scale", "in": "query", "description": "10000 expresses the ratio delta in basis points", "schema": { "type": "number", "format": "double", "exclusiveMinimum": 0, "maximum": 1000000000, "default": 10000 } }
        ],
        "responses": {
          "200": {
            "description": "Synthetic OHLC candles and close-series statistics",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ComparisonResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "422": { "$ref": "#/components/responses/NoOverlap" },
          "502": { "$ref": "#/components/responses/UpstreamError" },
          "503": { "$ref": "#/components/responses/Busy" },
          "504": { "$ref": "#/components/responses/UpstreamTimeout" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Venue": {
        "type": "string",
        "enum": ["binance_spot", "binance_perp", "bybit_spot", "bybit_perp", "hyperliquid_perp", "lighter_perp", "aster_perp", "ondo_perp", "mexc_spot", "mexc_perp", "okx_spot", "okx_perp"]
      },
      "Interval": {
        "type": "string",
        "enum": ["1m", "3m", "5m", "15m", "30m", "1h", "2h", "4h", "1d"]
      },
      "MarketSymbol": {
        "type": "string",
        "minLength": 1,
        "maxLength": 64,
        "pattern": "^[A-Za-z0-9_.:/@-]+$",
        "examples": ["BTCUSDT", "BTC_USDT", "BTC-USDT-SWAP", "BTC", "BTC-USD.P"]
      },
...
      "Market": {
        "type": "object",
        "additionalProperties": false,
        "required": ["symbol", "normalized_symbol", "base", "quote", "active"],
        "properties": {
          "symbol": { "$ref": "#/components/schemas/MarketSymbol" },
          "normalized_symbol": { "type": "string", "pattern": "^[A-Z0-9][A-Z0-9._:-]*(/[A-Z0-9][A-Z0-9._:-]*)?$", "example": "BTC/USDT" },
          "base": { "type": "string", "example": "BTC" },
          "quote": { "type": "string", "example": "USDT" },
          "active": { "type": "boolean" }
        }
      },
      "MarketsResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["venue", "query", "total", "cache_ttl_seconds", "data"],
        "properties": {
          "venue": { "$ref": "#/components/schemas/Venue" },
          "query": { "type": "string" },
          "total": { "type": "integer", "minimum": 0 },
          "cache_ttl_seconds": { "type": "integer", "const": 300 },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Market" } }
        }
      },
      "TickerListing": {
        "type": "object",
        "additionalProperties": false,
        "required": ["normalized_symbol", "symbol", "base", "quote", "venue", "venue_label", "market_type"],
        "properties": {
          "normalized_symbol": { "type": "string", "example": "BTC/USDT" },
          "symbol": { "$ref": "#/components/schemas/MarketSymbol" },
          "base": { "type": "string", "example": "BTC" },
          "quote": { "type": "string", "example": "USDT" },
          "venue": { "$ref": "#/components/schemas/Venue" },
          "venue_label": { "type": "string", "example": "Bybit Perpetual" },
          "market_type": { "type": "string", "enum": ["spot", "perpetual"] }
        }
      },
      "TickersResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["query", "venue", "total", "cache_ttl_seconds", "data"],
        "properties": {
          "query": { "type": "string" },
          "venue": { "oneOf": [{ "$ref": "#/components/schemas/Venue" }, { "type": "null" }] },
          "total": { "type": "integer", "minimum": 0 },
          "cache_ttl_seconds": { "type": "integer", "const": 300 },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/TickerListing" } }
        }
      },
      "Candle": {
        "type": "object",
        "additionalProperties": false,
        "required": ["time", "open", "high", "low", "close"],
        "properties": {
          "time": { "type": "integer", "format": "int64", "minimum": 0, "description": "Candle opening timestamp in Unix milliseconds" },
          "open": { "type": "number", "format": "double", "exclusiveMinimum": 0 },
          "high": { "type": "number", "format": "double", "exclusiveMinimum": 0 },
          "low": { "type": "number", "format": "double", "exclusiveMinimum": 0 },
          "close": { "type": "number", "format": "double", "exclusiveMinimum": 0 },
          "volume": { "type": "number", "format": "double", "minimum": 0 }
        }
      },
      "CandleResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["venue", "market", "interval", "count", "candles"],
        "properties": {
          "venue": { "$ref": "#/components/schemas/Venue" },
          "market": { "$ref": "#/components/schemas/MarketSymbol" },
          "interval": { "$ref": "#/components/schemas/Interval" },
          "count": { "type": "integer", "minimum": 0, "maximum": 1500 },
          "candles": { "type": "array", "maxItems": 1500, "items": { "$ref": "#/components/schemas/Candle" } }
        }
      },
      "ComparisonCandle": {
        "type": "object",
        "additionalProperties": false,
        "required": ["time", "open", "high", "low", "close", "left_close", "right_close"],
        "properties": {
          "time": { "type": "integer", "format": "int64", "minimum": 0 },
          "open": { "type": "number", "format": "double" },
          "high": { "type": "number", "format": "double" },
          "low": { "type": "number", "format": "double" },
          "close": { "type": "number", "format": "double" },
          "left_close": { "type": "number", "format": "double", "exclusiveMinimum": 0 },
          "right_close": { "type": "number", "format": "double", "exclusiveMinimum": 0 }
        }
      },
      "ComparisonStats": {
        "type": "object",
        "additionalProperties": false,
        "required": ["latest", "mean", "standard_deviation", "minimum", "maximum", "z_score"],
        "properties": {
          "latest": { "type": "number", "format": "double" },
          "mean": { "type": "number", "format": "double" },
          "standard_deviation": { "type": "number", "format": "double", "minimum": 0 },
          "minimum": { "type": "number", "format": "double" },
          "maximum": { "type": "number", "format": "double" },
          "z_score": { "type": "number", "format": "double" }
        }
      },
      "ComparisonResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["formula", "unit", "scale", "interval", "approximation", "candles", "stats", "matched_candles", "dropped_left", "dropped_right"],
        "properties": {
          "formula": { "type": "string", "example": "(bybit_perp:WLFIUSDT / mexc_perp:WLFI_USDT - 1) × 10000" },
          "unit": { "type": "string", "enum": ["bps", "scaled ratio delta"] },
          "scale": { "type": "number", "format": "double", "exclusiveMinimum": 0 },
          "interval": { "$ref": "#/components/schemas/Interval" },
          "approximation": { "type": "string" },
          "candles": { "type": "array", "minItems": 1, "maxItems": 1500, "items": { "$ref": "#/components/schemas/ComparisonCandle" } },
          "stats": { "$ref": "#/components/schemas/ComparisonStats" },
          "matched_candles": { "type": "integer", "minimum": 1, "maximum": 1500 },
          "dropped_left": { "type": "integer", "minimum": 0 },
          "dropped_right": { "type": "integer", "minimum": 0 }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "message"],
        "properties": {
          "code": { "type": "string", "enum": ["bad_request", "upstream_error", "upstream_timeout", "no_overlap", "busy", "internal_error"] },
          "message": { "type": "string" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": { "error": { "$ref": "#/components/schemas/ErrorDetail" } }
      }
    },
    "parameters": {
      "Venue": { "name": "venue", "in": "query", "required": true, "schema": { "$ref": "#/components/schemas/Venue" } },
      "Market": { "name": "market", "in": "query", "required": true, "schema": { "$ref": "#/components/schemas/MarketSymbol" }, "example": "BTCUSDT" },
      "Interval": { "name": "interval", "in": "query", "required": true, "schema": { "$ref": "#/components/schemas/Interval", "default": "1h" } },
      "From": { "name": "from", "in": "query", "required": true, "description": "Inclusive lower bound in Unix milliseconds", "schema": { "type": "integer", "format": "int64", "minimum": 0 } },
      "To": { "name": "to", "in": "query", "required": true, "description": "Exclusive upper bound in Unix milliseconds; must be later than from", "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
      "Limit": { "name": "limit", "in": "query", "description": "Maximum returned bars", "schema": { "type": "integer", "minimum": 1, "maximum": 1500, "default": 1000 } }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid or unsafe query parameters",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" }, "example": { "error": { "code": "bad_request", "message": "limit must be between 1 and 1500" } } } }
      },
      "NoOverlap": {
        "description": "The two normalized candle series have no identical opening timestamps",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" }, "example": { "error": { "code": "no_overlap", "message": "no candles overlap at identical timestamps" } } } }
      },
      "UpstreamError": {
        "description": "A venue rejected or failed the public market-data request",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" }, "example": { "error": { "code": "upstream_error", "message": "venue `example` rejected the request" } } } }
      },
      "UpstreamTimeout": {
        "description": "A venue exceeded the upstream deadline",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" }, "example": { "error": { "code": "upstream_timeout", "message": "venue `example` timed out" } } } }
      },
      "Busy": {
        "description": "The bounded upstream queue did not acquire capacity within two seconds",
        "headers": { "Retry-After": { "schema": { "type": "integer", "minimum": 1 }, "description": "Suggested retry delay in seconds" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" }, "example": { "error": { "code": "busy", "message": "service is busy; retry shortly" } } } }
      }
    }
  }
}